+++ backend/core/src/main/java/com/onair/hearit/core/domain/exception/DomainErrorCode.java
@@ -11,4 +11,5 @@ public enum DomainErrorCode {
     PLAYING_HISTORY_DOMAIN_ERROR,
     RECOMMEND_HEARIT_DOMAIN_ERROR,
     USERINFO_DOMAIN_ERROR,
+    REFRESH_TOKEN_DOMAIN_ERROR,
 }
