+++ backend/core/src/main/java/com/onair/hearit/core/domain/exception/RefreshTokenDomainException.java
@@ -1,0 +1,8 @@
+package com.onair.hearit.core.domain.exception;
+
+public class RefreshTokenDomainException extends DomainException {
+
+    public RefreshTokenDomainException(String message) {
+        super(DomainErrorCode.REFRESH_TOKEN_DOMAIN_ERROR, message);
+    }
+}
